backend/pyproject.toml
[tool.poetry]
name = "cybersecurity"
version = "0.1.0"
description = "cybersecurity project"
authors = ["houssam <houssamtouaj999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.22"
asgiref = "3.8.1"
django-admin-display = "1.3.0"
django-cors-headers = "4.7.0"
django-extensions = "4.1"
graphql-core = "3.2.6"
packaging = "25.0"
pillow = "11.2.1"
psycopg2-binary = "2.9.10"
PyJWT = "2.10.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
sqlparse = "0.5.3"
strawberry-django-jwt = "0.2.3"
strawberry-graphql = "0.271.1"
strawberry-graphql-django = "0.61.0"
typing_extensions = "4.14.0"
tzdata = "2025.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
